requirements.txt
tensorflow-cpu==2.17.1
streamlit
matplotlib
numpy
opencv-python
pillow
scikit-learn
pandas
seaborn